requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.19.2
requests==2.27.1
rsa==4.8
Telethon==1.24.0
urllib3==1.26.9
yarg==0.1.9
